Fix iOS build errors by consolidating AnyCodable implementation
This commit resolves all 21 iOS build errors related to AnyCodable:
- Created a single shared AnyCodable.swift file in Models/
- Removed duplicate AnyCodable declarations from Mission.swift
- Removed duplicate AnyCodable declarations from ToolUIModels.swift
- Added proper Hashable conformance to prevent type conformance errors
- Fixed ambiguous type lookup issues

The shared AnyCodable now properly implements:
- Codable protocol for JSON encoding/decoding
- Hashable protocol for use in dictionaries and sets
- Equatable protocol for comparison operations
- stringValue property for display formatting

Fixes:
- Invalid redeclaration of 'AnyCodable'
- 'AnyCodable' is ambiguous for type lookup
- Type 'String.Type' cannot conform to 'Hashable'
- No exact matches in call to instance method 'decode'
- Type conformance errors for Encodable/Decodable
