+++ ios_dashboard/SandboxedDashboard/Views/Components/ToolUI/ToolUIModels.swift
@@ -78,78 +78,6 @@ struct ToolUIOptionList: Codable {
     }
 }
 
-// MARK: - Any Codable Helper
-
-struct AnyCodable: Codable {
-    let value: Any
-    
-    init(_ value: Any) {
-        self.value = value
-    }
-    
-    init(from decoder: Decoder) throws {
-        let container = try decoder.singleValueContainer()
-        
-        if container.decodeNil() {
-            value = NSNull()
-        } else if let bool = try? container.decode(Bool.self) {
-            value = bool
-        } else if let int = try? container.decode(Int.self) {
-            value = int
-        } else if let double = try? container.decode(Double.self) {
-            value = double
-        } else if let string = try? container.decode(String.self) {
-            value = string
-        } else if let array = try? container.decode([AnyCodable].self) {
-            value = array.map { $0.value }
-        } else if let dict = try? container.decode([String: AnyCodable].self) {
-            value = dict.mapValues { $0.value }
-        } else {
-            throw DecodingError.dataCorruptedError(in: container, debugDescription: "Unable to decode value")
-        }
-    }
-    
-    func encode(to encoder: Encoder) throws {
-        var container = encoder.singleValueContainer()
-        
-        switch value {
-        case is NSNull:
-            try container.encodeNil()
-        case let bool as Bool:
-            try container.encode(bool)
-        case let int as Int:
-            try container.encode(int)
-        case let double as Double:
-            try container.encode(double)
-        case let string as String:
-            try container.encode(string)
-        case let array as [Any]:
-            try container.encode(array.map { AnyCodable($0) })
-        case let dict as [String: Any]:
-            try container.encode(dict.mapValues { AnyCodable($0) })
-        default:
-            try container.encode(String(describing: value))
-        }
-    }
-    
-    var stringValue: String {
-        switch value {
-        case is NSNull:
-            return "-"
-        case let bool as Bool:
-            return bool ? "Yes" : "No"
-        case let int as Int:
-            return int.formatted()
-        case let double as Double:
-            return double.formatted(.number.precision(.fractionLength(0...2)))
-        case let string as String:
-            return string
-        default:
-            return String(describing: value)
-        }
-    }
-}
-
 // MARK: - Progress Bar
 
 struct ToolUIProgress: Codable {
